requirements.txt
SQLAlchemy
flask_sqlalchemy
psycopg2
python-dotenv
flask-httpauth
pyjwt
marshmallow-sqlalchemy
flask-marshmallow
Flask-migrate
setuptools
flasgger
pandas
openpyxl